feat(draw_buf): distinguish between lv_image_dsc_t and lv_draw_buf_t
lv_image_dsc_t should only by used to describe a constant image
resource.

lv_draw_buf_t is used to describe an image in RAM.

Move _lv_image_buf_get_transformed_area to lv_draw_image.c
Adjust image flags to distinguish between flags exclusive to draw buf
