+++ demos/render/lv_demo_render.c
@@ -767,7 +767,7 @@ static void blend_mode_cb(lv_obj_t * parent)
     const char * cf_txt[] = {"RGB565", "RGB888.", "XRGB8888", "ARGB8888"};
     lv_color_format_t cf_values[] = {LV_COLOR_FORMAT_RGB565, LV_COLOR_FORMAT_RGB888, LV_COLOR_FORMAT_XRGB8888, LV_COLOR_FORMAT_ARGB8888};
     uint8_t * cf_bufs[] = {buf_rgb565, buf_rgb888, buf_xrgb8888, buf_argb8888};
-    static lv_image_dsc_t image_dscs[4];
+    static lv_draw_buf_t image_dscs[4];
 
     const char * mode_txt[] = {"Add.", "Sub.", "Mul."};
     lv_blend_mode_t mode_values[] = {LV_BLEND_MODE_ADDITIVE, LV_BLEND_MODE_SUBTRACTIVE, LV_BLEND_MODE_MULTIPLY};
@@ -787,7 +787,7 @@ static void blend_mode_cb(lv_obj_t * parent)
 
         lv_canvas_set_buffer(canvas, cf_bufs[cf], 36, 30, cf_values[cf]);
         create_blend_mode_image_buffer(canvas);
-        lv_img_dsc_t * img_src = lv_canvas_get_image(canvas);
+        lv_draw_buf_t * img_src = lv_canvas_get_draw_buf(canvas);
         image_dscs[cf] = *img_src;
 
         for(m = 0; m < 3; m++) {
