+++ lvgl.h
@@ -109,6 +109,7 @@ extern "C" {
 #include "src/layouts/lv_layout.h"
 
 #include "src/draw/lv_draw.h"
+#include "src/draw/lv_draw_buf.h"
 #include "src/draw/lv_draw_vector.h"
 
 #include "src/themes/lv_theme.h"
