+++ src/libs/gif/lv_gif.h
@@ -31,7 +31,7 @@ typedef struct {
     lv_image_t img;
     gd_GIF * gif;
     lv_timer_t * timer;
-    lv_image_dsc_t imgdsc;
+    lv_draw_buf_t imgdsc;
     uint32_t last_call;
 } lv_gif_t;
